fix(ApplicationTest): ApplicationTest 변경 내용 취소
